# src/FinancialPortfolioAPI.API/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Financial Portfolio API",
    "version": "1.0.0",
    "description": "SQLite or SQL Server test-server API with transactional settlement, soft deletion, audit history and committed changes. USD and fixed demo prices. See /docs/guide."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKey": []
    }
  ],
  "paths": {
...
    "/api/v1/portfolio/{id}": {
      "get": {
        "summary": "Read a portfolio without changing it",
        "operationId": "getapivportfolioid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace portfolio metadata with a version check",
        "operationId": "putapivportfolioid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePortfolioRequest"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Soft-delete an empty portfolio and retain history",
        "operationId": "deletePortfolio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expectedVersion",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or portfolio is not empty"
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transaction/portfolio/{portfolioId}": {
      "get": {
        "summary": "List portfolio transactions",
        "operationId": "getapivtransactionportfolioportfolioId",
        "parameters": [
          {
            "name": "portfolioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create pending transaction; identical reference and payload returns original",
        "operationId": "postapivtransactionportfolioportfolioId",
        "parameters": [
          {
            "name": "portfolioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "Location": {
                "description": "Resource URL",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/transaction/{id}": {
      "get": {
        "summary": "Read a transaction",
        "operationId": "getapivtransactionid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transaction/{id}/status": {
      "patch": {
        "summary": "Advance transaction lifecycle; settlement applies balances atomically",
        "operationId": "patchapivtransactionidstatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/changes": {
      "get": {
        "summary": "Poll committed resource changes in sequence order",
        "operationId": "getapivchanges",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prices": {
      "get": {
        "summary": "List fixed demo quotes, not live prices",
        "operationId": "getapivprices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prices"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
      "RiskProfile": {
        "type": "string",
        "enum": [
          "Conservative",
          "Moderate",
          "Aggressive",
          "VeryAggressive"
        ]
      },
      "PortfolioStatus": {
        "type": "string",
        "enum": [
          "Active",
          "Inactive",
          "PendingReview",
          "Restricted",
          "Closed"
        ]
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "Buy",
          "Sell",
          "Dividend",
          "Deposit",
          "Withdrawal"
        ]
      },
      "TransactionStatus": {
        "type": "string",
        "enum": [
          "Pending",
          "Executed",
          "Settled",
          "Failed",
          "Cancelled"
        ]
      },
...
      "UpdatePortfolioRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "riskProfile": {
            "$ref": "#/components/schemas/RiskProfile"
          },
          "status": {
            "$ref": "#/components/schemas/PortfolioStatus"
          }
        },
        "required": [
          "expectedVersion",
          "name",
          "riskProfile",
          "status"
        ]
      },
...
      "StatusRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          }
        },
        "required": [
          "expectedVersion",
          "status"
        ]
      },
      "CreateTransactionRequest": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "referenceNumber": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "symbol": {
            "type": "string",
            "maxLength": 20,
            "nullable": true
          },
          "shares": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 1000000000,
            "default": 0
          },
          "price": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 1000000000,
            "default": 0
          },
          "amount": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 1000000000000,
            "default": 0
          },
          "commission": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 1000000,
            "default": 0
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "default": ""
          }
        },
        "required": [
          "type",
          "referenceNumber"
        ]
      },
      "Holding": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "assetClass": {
            "type": "string",
            "enum": [
              "Equity",
              "FixedIncome",
              "Alternative",
              "Cash",
              "Commodity",
              "RealEstate",
              "Cryptocurrency"
            ]
          },
          "shares": {
            "type": "number",
            "format": "double"
          },
          "currentPrice": {
            "type": "number",
            "format": "double"
          },
          "costBasis": {
            "type": "number",
            "format": "double"
          },
          "lastPriceUpdate": {
            "type": "string",
            "format": "date-time"
          },
          "marketValue": {
            "type": "number",
            "format": "double"
          },
          "gainLoss": {
            "type": "number",
            "format": "double"
          },
          "gainLossPercentage": {
            "type": "number",
            "format": "double"
          },
          "portfolioId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "symbol",
          "companyName",
          "assetClass",
          "shares",
          "currentPrice",
          "costBasis",
          "lastPriceUpdate",
          "marketValue",
          "gainLoss",
          "gainLossPercentage"
        ]
      },
      "Portfolio": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "clientName": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "cashBalance": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/PortfolioStatus"
          },
          "riskProfile": {
            "$ref": "#/components/schemas/RiskProfile"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "holdings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Holding"
            }
          },
          "totalValue": {
            "type": "number",
            "format": "double"
          },
          "unrealizedGainLoss": {
            "type": "number",
            "format": "double"
          },
          "allocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            }
          },
          "isDeleted": {
            "type": "boolean"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "clientId",
          "clientName",
          "currency",
          "cashBalance",
          "status",
          "riskProfile",
          "version",
          "createdAt",
          "lastUpdatedAt",
          "holdings",
          "totalValue",
          "unrealizedGainLoss",
          "allocation"
        ]
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "portfolioId": {
            "type": "string",
            "format": "uuid"
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "shares": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "commission": {
            "type": "number",
            "format": "double"
          },
          "netAmount": {
            "type": "number",
            "format": "double"
          },
          "referenceNumber": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          },
          "settlementDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "id",
          "portfolioId",
          "symbol",
          "type",
          "shares",
          "price",
          "amount",
          "commission",
          "netAmount",
          "referenceNumber",
          "notes",
          "status",
          "transactionDate",
          "settlementDate",
          "version"
        ]
      },
      "Change": {
        "type": "object",
        "properties": {
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "committedAt": {
            "type": "string",
            "format": "date-time"
          },
          "action": {
            "type": "string",
            "description": "Updated or Deleted. Deleted portfolios subsequently return 404."
          }
        },
        "required": [
          "sequence",
          "entityType",
          "entityId",
          "version",
          "committedAt"
        ]
      },
      "ChangePage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Change"
            }
          },
          "nextCursor": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "Quote": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "symbol",
          "companyName",
          "price"
        ]
      },
      "Prices": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quote"
            }
          }
        },
        "required": [
          "source",
          "currency",
          "items"
        ]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "title",
          "status"
        ]
      },
...
      "TransactionPage": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "items"
        ]
      },
...
    }
  }
}